mcp-server/database.py
"""Database connection and configuration for MIVA Academic MCP Server."""

import os
from typing import Any, Dict, List, Optional
from contextlib import asynccontextmanager
from dotenv import load_dotenv

# Load environment variables
load_dotenv()


class DatabaseConfig:
    """Database configuration class."""
    
    def __init__(self):
        self.database_url = os.getenv('DATABASE_URL')
        self.host = os.getenv('DB_HOST', 'localhost')
        self.port = int(os.getenv('DB_PORT', '5432'))
        self.database = os.getenv('DB_NAME', 'miva_academic')
        self.user = os.getenv('DB_USER', 'postgres')
        self.password = os.getenv('DB_PASSWORD', '')


class AcademicRepository:
    """Repository class for academic database operations."""
    
    def __init__(self, config: DatabaseConfig):
        self.config = config
        # TODO: Initialize database connection pool
        # This will be implemented when psycopg2 is available
        self._connection = None
    
    async def get_course_materials(
        self,
        course_code: str,
        student_id: str,
        week_number: Optional[int] = None,
        material_type: Optional[str] = None
    ) -> Dict[str, Any]:
        """Get course materials for a student."""
        # TODO: Implement actual database query
        # This is a placeholder that returns sample data
        
        # Verify student enrollment
        is_enrolled = await self._check_enrollment(student_id, course_code)
        if not is_enrolled:
            return {
                "error": "Access denied: Student not enrolled in this course",
                "course_code": course_code,
                "student_id": student_id
            }
        
        # Sample data structure
        materials = [
            {
                "id": "1",
                "week_number": 1,
                "material_type": "lecture",
                "title": f"{course_code} Week 1 Lecture Notes",
                "description": "Introduction to course concepts",
                "file_url": "/materials/cs101_week1_lecture.pdf",
                "upload_date": "2024-01-15"
            },
            {
                "id": "2", 
                "week_number": 1,
                "material_type": "reading",
                "title": "Required Reading Chapter 1",
                "description": "Fundamental concepts overview",
                "file_url": "/materials/cs101_chapter1.pdf",
                "upload_date": "2024-01-15"
            }
        ]
        
        # Apply filters
        if week_number:
            materials = [m for m in materials if m["week_number"] == week_number]
        if material_type:
            materials = [m for m in materials if m["material_type"] == material_type]
        
        return {
            "course_code": course_code,
            "materials": materials,
            "total_count": len(materials)
        }
    
    async def get_upcoming_assignments(
        self,
        student_id: str,
        days_ahead: int = 7,
        course_code: Optional[str] = None,
        include_completed: bool = False
    ) -> Dict[str, Any]:
        """Get upcoming assignments for a student."""
        # TODO: Implement actual database query
        
        # Sample assignment data
        assignments = [
            {
                "id": "1",
                "course_code": "CS101",
                "title": "Programming Assignment 1",
                "description": "Implement basic algorithms",
                "due_date": "2024-02-01",
                "due_time": "23:59",
                "points_possible": 100,
                "urgency": "soon",
                "days_until_due": 2,
                "status": "pending"
            },
            {
                "id": "2", 
                "course_code": "MATH201",
                "title": "Calculus Problem Set 3",
                "description": "Solve integration problems",
                "due_date": "2024-02-05",
                "due_time": "23:59", 
                "points_possible": 75,
                "urgency": "later",
                "days_until_due": 6,
                "status": "pending"
            }
        ]
        
        # Apply filters
        if course_code:
            assignments = [a for a in assignments if a["course_code"] == course_code.upper()]
        if not include_completed:
            assignments = [a for a in assignments if a["status"] != "completed"]
        
        # Filter by days ahead
        assignments = [a for a in assignments if a["days_until_due"] <= days_ahead]
        
        return {
            "student_id": student_id,
            "assignments": assignments,
            "total_count": len(assignments)
        }
    
    async def get_course_info(self, course_code: str) -> Dict[str, Any]:
        """Get detailed course information."""
        # TODO: Implement actual database query
        
        return {
            "course_code": course_code,
            "course_name": f"Sample Course {course_code}",
            "description": "Course description placeholder",
            "credits": 3,
            "instructor": "Dr. Sample Professor",
            "semester": "Fall 2024",
            "meeting_times": "MWF 10:00-11:00 AM"
        }
    
    async def _check_enrollment(self, student_id: str, course_code: str) -> bool:
        """Check if student is enrolled in course."""
        # TODO: Implement actual enrollment check
        # For now, return True as placeholder
        return True
    
    async def close(self):
        """Close database connections."""
        if self._connection:
            # TODO: Implement connection cleanup
            pass


# Global repository instance
db_config = DatabaseConfig()
academic_repo = AcademicRepository(db_config)